+++ src/btclib/ecc/ecies.py
@@ -77,10 +77,17 @@ from dataclasses import dataclass
 from hashlib import sha256, sha512
 
 from btclib.alias import CipherF, Integer, Octets, String
-from btclib.curves import bytes_from_point, mult, scalar_from_prv_key, secp256k1
+from btclib.curves import (
+    PubKey,
+    bytes_from_point,
+    mult,
+    point_from_pub_key,
+    scalar_from_prv_key,
+    secp256k1,
+)
 from btclib.curves.sec_point import _mult_sec_var
 from btclib.exceptions import BTClibRuntimeError, BTClibValueError
-from btclib.to_pub_key import PubKey, point_from_pub_key, pub_keyinfo_from_pub_key
+from btclib.to_pub_key import pub_keyinfo_from_pub_key
 from btclib.utils import assert_type, bytes_from_octets, str_from_string
 
 __all__ = [